Don't need an id.
Is infered directly from the context in which a post lives. 

ie posts live under a persons directory so the path of the document is 
information the doucment can use to determine identity. 

/{phone_numper}/posts

posts are unique to a user and so user id and created_at function as a 
globally unique id. so there is no reason to duplicate that information 
with every individual post. 

anyways, that's my reasoning for removing explicit identity.