Stick with the first found value for an itemprop
this is a fix for parent resources taking on the properties of their 
last child 

This is not an ideal fix as I would rather have a query that gets all 
the properties of an Item and respects the properties of any child items 
something like

:not([itemscope]) itemscope

though that query will not work.

created_at is especially vlunerable to this bug.