Show the done button on screens bigger then phone
I kind of hate this. Ipads have this same resolution. 

so there is a done button ob a screeb that already has a done button. 

lame.

but I want the done buttton to show up on computors. 

there is no current selector for wether the user is being shown a 
keyboard. 

so I have to half ass this and just show the done button borking Ipads 
in favor of a done button becuase computors need a done button for the 
mouse people. 

though computer people should just use the tabe button.