Renaming posts to statements
each statment can awnser the 5 basic journalistic questions. 

The better the system's perception the more we will be able to provide 
clarity about what people are talking about. 


like 

so for that reason I am going to include why and where as part of 
statments. Where is geolocation plus a selector.

why is context. Can you wrap a post in it's context. Just posted a 
flyer, just lost their job to a pandemic. what does the system percieve 
that gives this post context. 

on machine Context