Extend the animation effect to all of  the menu items.
This is intended to first focus the user on their avatar picture which 
should be kind of awesome (cause vectors are rad)

I also like how there are buttons at the corners. and a cluster of 4 
buttons  at the bottom with just a tiny bit of scrolling. everything is 
right there yet it looks big on a small phone.