Fix for relationship status not rendering
This is a troubling fix. why would a signed_in check require to start 
off it's status as true in order for the component to know about it's 
relationships. 

if signed_in is initialized true then relationship status is passed 
along

if it is false than the relations become an empty list. 

not sure what's going on here but I don't like it