Pass elements that are synced to the save command
This was a nasty little bug. because I wasn't explicitly passing the 
elements i wanted to save the save method in Local.js was doing a lookup 
for the elements and in this ONE case it was on the page twice and 
saving the wrong one. 

use the sync ref to make sure my lookups are contextualized