Don't need raw text content. item curently handle svg content reasonably
I will wrap the animation in a symbol a nd it will be treated like a 
property of a poster. 

This is why you keep your business loging as small as possible. I just 
had to read it real quick to remember take advantage of previous 
thinking. 

"oh yeah, just use a symbol tag for animations"