Setup javascript in vscode
This is me giving typescript the best possible chance. Typescript is 
developemnt typing it's not production typing. Strange. and you have to 
do all this setup to see what it gets you. 

I'm still not convinced, but I will damn sure know all about it before 
it dismiss it for my own code