fix: Rollback creating subdirectories for posters
We are gonna try and get it to work with the standard google list for storage objects. This approacha may not work but would reduce the complexity. I just have to figure out how to get the result set in reverse lexical order which is what gives me newer numbers first.

fix: Can posters all the way back to the first poster. (we get validation warnings but that is chill for how old these posters are)